Move: 검색 페이지 스켈렙톤 파일명 수정
