+++ src/app/(navbar)/search/_component/SkeletonItem.tsx
